code/ai-podcast-ai-agent/requirements.txt
crewai
crewai-tools
# langchain_ollama>=0.0.3
gradio
elevenlabs
python-dotenv
firecrawl